#25 구현 완료, #27 구현 진행중
