+++ app/src/main/java/com/gunjun/android/personalproject/InstagramActivity.java
@@ -1,0 +1,13 @@
+package com.gunjun.android.personalproject;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class InstagramActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_instagram);
+    }
+}
