+++ app/src/main/res/layout/activity_instagram.xml
@@ -1,0 +1,16 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context="com.gunjun.android.personalproject.fragment.YoutubeFragment">
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/youtube_list"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:padding="@dimen/activity_horizontal_margin"
+        android:paddingTop="?attr/actionBarSize"
+        tools:context=".YoutubeActivity"
+        tools:listitem="@layout/youtube_list" />
+
+</LinearLayout>
